Update the test for controller
